Responsive single coin page and centralize style files
