+++ src/components/CoinInfo/CoinInfo.js
@@ -29,7 +29,7 @@ function CoinInfo(props) {
 
   return (
     <TableRow>
-      <TableCol rank={true} xs={4} sm={4} md={3} lg={3} xl={2} xxl={2}>
+      <TableCol rank={true} xs={3} sm={4} md={3} lg={3} xl={2} xxl={2}>
         <StyledLink
           key={coin}
           to={`/coins/${coin.id}?currency=${currency}`}
@@ -38,7 +38,7 @@ function CoinInfo(props) {
           {coin.marketCapRank}
         </StyledLink>
       </TableCol>
-      <TableCol xs={6} sm={4} md={4} lg={4} xl={4} xxl={4}>
+      <TableCol xs={6} sm={4} md={6} lg={4} xl={4} xxl={4}>
         <StyledLink
           key={coin}
           to={`/coins/${coin.id}?currency=${currency}`}
@@ -65,7 +65,7 @@ function CoinInfo(props) {
           %
         </StyledPercentage>
       </TableCol>
-      <TableCol xs={4} sm={4} md={2} lg={2} xl={2} xxl={2}>
+      <TableCol xs={5} sm={4} md={2} lg={2} xl={2} xxl={2}>
         <StyledPercentage increase={increase24h}>
           <IncreaseArrow src={increase24h ? increaseArrow : decreaseArrow} />
           {coin.priceChangePercentage24HInCurrency &&
@@ -81,14 +81,14 @@ function CoinInfo(props) {
           %
         </StyledPercentage>
       </TableCol>
-      <TableCol xs={0} sm={0} md={4} lg={4} xl={3} xxl={3}>
+      <TableCol xs={0} sm={0} md={6} lg={4} xl={3} xxl={3}>
         <MarketInfo
           numerator={coin.totalVolume}
           denominator={coin.marketCap}
           currency={currency}
         />
       </TableCol>
-      <TableCol xs={0} sm={0} md={4} lg={4} xl={3} xxl={3}>
+      <TableCol xs={0} sm={0} md={0} lg={4} xl={3} xxl={3}>
         <MarketInfo
           numerator={coin.circulatingSupply}
           denominator={coin.totalSupply}
